moved CutPlaneSpec to butt_joint
